Update portage tree

+++ sys-apps/groff/groff-1.22.2.ebuild
@@ -1,22 +1,22 @@
-# Copyright 1999-2014 Gentoo Foundation
+# Copyright 1999-2017 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/sys-apps/groff/groff-1.22.2.ebuild,v 1.13 2014/01/18 19:29:26 vapier Exp $
+# $Id$
 
 EAPI="4"
 
 inherit autotools eutils toolchain-funcs
 
 DESCRIPTION="Text formatter used for man pages"
-HOMEPAGE="http://www.gnu.org/software/groff/groff.html"
+HOMEPAGE="https://www.gnu.org/software/groff/groff.html"
 SRC_URI="mirror://gnu/groff/${P}.tar.gz
-	linguas_ja? ( http://dev.gentoo.org/~naota/patch/${PN}-1.22.1-japanese.patch.bz2 )"
+	l10n_ja? ( https://dev.gentoo.org/~naota/patch/${PN}-1.22.1-japanese.patch.bz2 )"
 
 LICENSE="GPL-2"
 SLOT="0"
-KEYWORDS="alpha amd64 arm arm64 hppa ia64 m68k ~mips ppc ppc64 s390 sh sparc x86 ~ppc-aix ~amd64-fbsd ~sparc-fbsd ~x86-fbsd ~x64-freebsd ~x86-freebsd ~ia64-hpux ~x86-interix ~amd64-linux ~arm-linux ~ia64-linux ~x86-linux ~ppc-macos ~x64-macos ~x86-macos ~m68k-mint ~sparc-solaris ~sparc64-solaris ~x64-solaris ~x86-solaris"
-IUSE="examples X linguas_ja"
+KEYWORDS="alpha amd64 arm arm64 hppa ia64 m68k ~mips ppc ppc64 s390 sh sparc x86 ~ppc-aix ~amd64-fbsd ~sparc-fbsd ~x86-fbsd ~amd64-linux ~arm-linux ~x86-linux ~ppc-macos ~x64-macos ~x86-macos ~m68k-mint ~sparc-solaris ~sparc64-solaris ~x64-solaris ~x86-solaris"
+IUSE="examples X l10n_ja"
 
-RDEPEND=">=sys-apps/texinfo-4.7-r1
+RDEPEND="
 	X? (
 		x11-libs/libX11
 		x11-libs/libXt
@@ -26,7 +26,7 @@ RDEPEND=">=sys-apps/texinfo-4.7-r1
 		x11-libs/libICE
 	)"
 DEPEND="${RDEPEND}
-	linguas_ja? ( virtual/yacc )"
+	l10n_ja? ( virtual/yacc )"
 
 DOCS=( BUG-REPORT ChangeLog MORE.STUFF NEWS PROBLEMS PROJECTS README REVISION TODO VERSION )
 
@@ -53,7 +53,7 @@ src_prepare() {
 	.ds default-operating-system Gentoo${pfx}/${KERNEL}
 	EOF
 
-	if use linguas_ja ; then
+	if use l10n_ja ; then
 		epatch "${WORKDIR}"/${PN}-1.22.1-japanese.patch #255292 #350534 #450796
 		eautoconf
 		eautoheader
@@ -69,7 +69,7 @@ src_configure() {
 		--with-appresdir="${EPREFIX}"/usr/share/X11/app-defaults \
 		--docdir="${EPREFIX}"/usr/share/doc/${PF} \
 		$(use_with X x) \
-		$(usex linguas_ja '--enable-japanese' '')
+		$(usex l10n_ja '--enable-japanese' '')
 }
 
 src_compile() {
@@ -84,7 +84,12 @@ src_install() {
 	dosym tbl /usr/bin/gtbl
 
 	if ! use examples ; then
+		# The pdf files might not be generated if ghostscript is unavailable. #602020
+		local pdf="${ED}/usr/share/doc/${PF}/examples/mom/mom-pdf.pdf"
+		if [[ -e ${pdf} ]] ; then
+			# Keep mom-pdf.pdf since it's more of a manual than an example. #454196 #516732
+			mv "${pdf}" "${ED}"/usr/share/doc/${PF}/pdf/ || die
+		fi
 		rm -rf "${ED}"/usr/share/doc/${PF}/examples
-		rm -f "${ED}"/usr/share/doc/${PF}/pdf/mom-pdf.pdf
 	fi
 }
